Finalized auth functions

+++ src/components/Layout/Header.tsx
@@ -11,15 +11,38 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
+import { GenericResponse } from "@/lib/types";
+import axios, { AxiosRequestConfig } from "axios";
 import { useNavigate } from 'react-router-dom'
 
 const Header = () => {
   const navigate = useNavigate();
   const usermail = localStorage.getItem('email');
-  const logout = () => {
-    localStorage.clear();
+  const logout = async () => {
+    const config: AxiosRequestConfig = {
+      url: "auth/logout",
+      method: "POST",
+      headers: {
+        Authorization: `Bearer ${localStorage.getItem('token')}`
+      }
+    }
+    try {
+      const res: GenericResponse = (await axios.request(config)).data;
+      if (res.success) {
+        localStorage.clear();
+        navigate("/auth/login");
+        return;
+      }
+      console.log(res);
+      navigate("/auth/login");
+    } catch (error) {
+      console.log(error);
+      localStorage.clear();
+      navigate("/auth/login");
+    }
     navigate("/auth/login");
   }
+
   const profile = () => navigate("/profile");
 
   return (
